+++ lib/rspec/matchers/built_in/raise_error.rb
@@ -4,6 +4,7 @@ module RSpec
       # @api private
       # Provides the implementation for `raise_error`.
       # Not intended to be instantiated directly.
+      # rubocop:disable ClassLength
       class RaiseError
         include Composable
 
@@ -63,6 +64,10 @@ module RSpec
 
         # @private
         def supports_block_expectations?
+          true
+        end
+
+        def expects_call_stack_jump?
           true
         end
 
@@ -163,6 +168,7 @@ module RSpec
           raise "`expect { }.to raise_error(message).with_message(message)` is not valid. The matcher only allows the expected message to be specified once"
         end
       end
+      # rubocop:enable ClassLength
     end
   end
 end