+++ lib/rspec/matchers/built_in/throw_symbol.rb
@@ -94,6 +94,10 @@ module RSpec
           true
         end
 
+        def expects_call_stack_jump?
+          true
+        end
+
       private
 
         def actual_result