+++ Changelog.md
@@ -6,6 +6,12 @@ Bug Fixes:
 * Rename private `LegacyMacherAdapter` constant to `LegacyMatcherAdapter`
   to fix typo. (Abdelkader Boudih, #563)
 
+Enhancements:
+
+* Extended compound matcher support to block matchers, for cases like:
+  `expect { ... }.to change { x }.to(3).and change { y }.to(4)`. (Myron
+  Marston, #567)
+
 ### 3.0.2 / 2014-06-19
 [Full Changelog](http://github.com/rspec/rspec-expectations/compare/v3.0.1...v3.0.2)